--- packages/clerk-js/src/utils/cookies/index.ts
@@ -1,1 +1,0 @@
-export * from './handler';
