+++ packages/shared/src/index.ts
@@ -30,10 +30,4 @@ export * from './proxy';
 export * from './underscore';
 export * from './url';
 export { createWorkerTimers } from './workerTimers';
-export {
-  DEV_BROWSER_SSO_JWT_PARAMETER,
-  DEV_BROWSER_JWT_MARKER,
-  getDevBrowserJWTFromURL,
-  setDevBrowserJWTInURL,
-  getDevBrowserJWTFromResponse,
-} from './devBrowser';
+export { DEV_BROWSER_JWT_KEY, getDevBrowserJWTFromURL, setDevBrowserJWTInURL } from './devBrowser';
